extract ScanTree
